requirements.txt
python-telegram-bot[job-queue]>=20.0,<21.0
SQLAlchemy>=1.4,<2.1
python-dotenv>=0.20,<1.1